refactor(plugins): split runtime deps planner
